Completely working with both compile and no compile modes.
Had to drop the jam less dependency because we need more control of how
less is run depending if it is a build or not. The included less and
less-rhino files are slightly altered from the 1.3.1 files to make it
work.

Bumped the version of the package to match the less source.